Added Integration tests

+++ src/api/v1/services/MealService.ts
@@ -1,11 +1,11 @@
-import mongoose, { ObjectId, UpdateQuery, FilterQuery } from 'mongoose';
+import mongoose, { UpdateQuery, FilterQuery } from 'mongoose';
 
 import { IMeal } from '../interfaces/models';
 import { PromiseHandler } from '../interfaces/types';
-import MealRepository from '../repositories/MealRepository';
+import { MealRepository, MealRepositoryType } from '../repositories';
 
 export default class MealService {
-  private readonly _repo: MealRepository = new MealRepository();
+  private readonly _repo: MealRepositoryType = new MealRepository();
 
   async getAll(): PromiseHandler<IMeal[]> {
     try {
@@ -17,6 +17,8 @@ export default class MealService {
   }
 
   async getById(id: string): PromiseHandler<IMeal> {
+    if (id.length !== 24)
+      return [undefined, new Error('Id must be 12 chars long')];
     const objectId: mongoose.Types.ObjectId = this.createIdFromString(id);
     try {
       const meal: IMeal | null = await this._repo.findById(objectId);
@@ -46,10 +48,12 @@ export default class MealService {
   }
 
   async update(id: string, data: UpdateQuery<IMeal>): PromiseHandler<IMeal> {
+    if (id.length !== 24)
+      return [undefined, new Error('Id must be 12 chars long')];
     const objectId: mongoose.Types.ObjectId = this.createIdFromString(id);
     try {
       const meal: IMeal | null = await this._repo.update(objectId, data);
-      if (!meal) return [undefined, new Error('No meal found')];
+      if (!meal) return [undefined, new Error('Invalid Id')];
       return [meal, undefined];
     } catch (err) {
       return [undefined, err as Error];
@@ -57,6 +61,8 @@ export default class MealService {
   }
 
   async delete(id: string): PromiseHandler<IMeal> {
+    if (id.length !== 24)
+      return [undefined, new Error('Id must be 12 chars long')];
     const objectId: mongoose.Types.ObjectId = this.createIdFromString(id);
     try {
       const meal: IMeal | null = await this._repo.delete(objectId);
